+++ src/data_provider/src/packages/packageLinuxParserHelper.h
@@ -71,7 +71,7 @@ namespace PackageLinuxHelper
             std::string version {UNKNOWN_VALUE};
             std::string vendor {UNKNOWN_VALUE};
             std::string description {UNKNOWN_VALUE};
-            int64_t size { 0 };
+            int size                 { 0 };
 
             auto it{info.find("Priority")};
 
@@ -91,7 +91,7 @@ namespace PackageLinuxHelper
 
             if (it != info.end())
             {
-                size = stoll(it->second) * 1024;
+                size = stol(it->second) * 1024;
             }
 
             it = info.find("Multi-Arch");
@@ -163,7 +163,7 @@ namespace PackageLinuxHelper
         std::string vendor       { UNKNOWN_VALUE };
         std::string install_time { UNKNOWN_VALUE };
         std::string description  { UNKNOWN_VALUE };
-        int64_t     size         { 0 };
+        int         size         { 0 };
         bool        hasName      { false };
         bool        hasVersion   { false };
 
@@ -224,7 +224,7 @@ namespace PackageLinuxHelper
 
             if (data.is_number())
             {
-                size = data.get<int64_t>();
+                size = data.get<int>();
             }
             else if (data.is_string())
             {
@@ -234,7 +234,7 @@ namespace PackageLinuxHelper
                 {
                     try
                     {
-                        size = std::stoll( stringData );
+                        size = std::stoi( stringData );
                     }
                     catch (const std::exception& e)
                     {
